+++ docs/source/changelog.rst
@@ -8,6 +8,7 @@ Paver Changelog
 * Ability to share command line options between tasks (github issue `#7 <https://github.com/paver/paver/issues/issue/7>`_)
 * Flush after print (github issue `#17 <https://github.com/paver/paver/issues/issue/17>`_, thanks to Honza Kral)
 * Minilib is now compatible with zipimport (github issue `#19 <https://github.com/paver/paver/issues/issue/19>`_, thanks to Piet Delport)
+* Auto task is now properly not called when target task is decorated with no_auto (github issue `#4 <https://github.com/paver/paver/issues/issue/24>`_)
 
 
 1.0.4 (January 16, 2011)